rename e2e directory to kube_e2e
