+++ test/kube_e2e/containers/event-emitter/main.go
