+++ test/kube_e2e/containers/grpc-test-service/Dockerfile
