+++ test/kube_e2e/containers/grpc-test-service/main.go
@@ -9,8 +9,8 @@ import (
 	"github.com/grpc-ecosystem/go-grpc-middleware"
 	"github.com/grpc-ecosystem/go-grpc-middleware/logging/zap"
 	"github.com/grpc-ecosystem/go-grpc-middleware/tags"
-	"github.com/solo-io/gloo/test/e2e/containers/grpc-test-service/bookstore/protos"
-	"github.com/solo-io/gloo/test/e2e/containers/grpc-test-service/server"
+	"github.com/solo-io/gloo/test/kube_e2e/containers/grpc-test-service/bookstore/protos"
+	"github.com/solo-io/gloo/test/kube_e2e/containers/grpc-test-service/server"
 	"go.uber.org/zap"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/reflection"
