+++ test/kube_e2e/containers/helloservice/Dockerfile
