+++ test/kube_e2e/containers/testrunner/Dockerfile
