+++ test/kube_e2e/containers/upstream-for-events/Dockerfile
