+++ test/kube_e2e/function_discovery_test.go
@@ -1,4 +1,4 @@
-package e2e
+package kube_e2e
 
 import (
 	. "github.com/onsi/ginkgo"
