+++ test/kube_e2e/grpc_test.go
@@ -1,4 +1,4 @@
-package e2e
+package kube_e2e
 
 import (
 	"time"
