+++ test/kube_e2e/kube_resources/helm-values.yaml.tmpl
