+++ test/kube_e2e/kube_resources/testing-resources.yaml.tmpl
