+++ test/kube_e2e/kube_upstream_discovery_test.go
@@ -1,4 +1,4 @@
-package e2e
+package kube_e2e
 
 import (
 	"time"
@@ -7,30 +7,16 @@ import (
 	"github.com/pborman/uuid"
 	"github.com/solo-io/gloo/pkg/api/types/v1"
 	. "github.com/solo-io/gloo/test/helpers"
-	"github.com/solo-io/gloo/pkg/coreplugins/service"
 )
 
-var _ = Describe("Core Service Plugin", func() {
-	const helloService = "helloservice"
-	const servicePort = 8080
-	Context("creating service upstream and a vhost with a single route to it", func() {
+var _ = Describe("Kube Upstream Discovery", func() {
+	generatedUpstreamName := namespace + "-helloservice-8080"
+	Context("creating a vhost with a route to an upstream generated by"+
+		" kube-upstream-discovery", func() {
 		randomPath := "/" + uuid.New()
 		vhostName := "one-route"
 		BeforeEach(func() {
-			_, err := gloo.V1().Upstreams().Create(&v1.Upstream{
-				Name: helloService,
-				Type: service.UpstreamTypeService,
-				Spec: service.EncodeUpstreamSpec(service.UpstreamSpec{
-					Hosts: []service.Host{
-						{
-							Addr: helloService,
-							Port: servicePort,
-						},
-					},
-				}),
-			})
-			Must(err)
-			_, err = gloo.V1().VirtualHosts().Create(&v1.VirtualHost{
+			_, err := gloo.V1().VirtualHosts().Create(&v1.VirtualHost{
 				Name: vhostName,
 				Routes: []*v1.Route{{
 					Matcher: &v1.Route_RequestMatcher{
@@ -44,7 +30,7 @@ var _ = Describe("Core Service Plugin", func() {
 					SingleDestination: &v1.Destination{
 						DestinationType: &v1.Destination_Upstream{
 							Upstream: &v1.UpstreamDestination{
-								Name: helloService,
+								Name: generatedUpstreamName,
 							},
 						},
 					},
@@ -53,14 +39,11 @@ var _ = Describe("Core Service Plugin", func() {
 			Must(err)
 		})
 		AfterEach(func() {
-			gloo.V1().Upstreams().Delete(helloService)
+			//gloo.V1().Upstreams().Delete(generatedUpstreamName)
 			gloo.V1().VirtualHosts().Delete(vhostName)
 		})
 		It("should configure envoy with a 200 OK route (backed by helloservice)", func() {
 			curlEventuallyShouldRespond(curlOpts{path: randomPath}, "< HTTP/1.1 200", time.Minute*5)
-		})
-		It("POST should be 404", func() {
-			curlEventuallyShouldRespond(curlOpts{path: randomPath, method: "POST"}, "< HTTP/1.1 404", time.Minute*5)
 		})
 	})
 })
