+++ test/kube_e2e/kubernetes_ingress_controller_test.go
@@ -1,4 +1,4 @@
-package e2e
+package kube_e2e
 
 import (
 	"time"
@@ -62,7 +62,7 @@ var _ = Describe("Kubernetes Ingress Controller", func() {
 					return nil, err
 				}
 				return updated.Status.LoadBalancer.Ingress, nil
-			}, "1m","5s").Should(Equal(expectedLbStatus))
+			}, "1m", "5s").Should(Equal(expectedLbStatus))
 		})
 	})
 	Context("creating a kubernetes ingress with multiple rules", func() {
