+++ test/kube_e2e/kubernetes_suite_test.go
@@ -1,4 +1,4 @@
-package e2e
+package kube_e2e
 
 import (
 	. "github.com/onsi/ginkgo"
