+++ test/kube_e2e/kubernetes_upstream_plugin_test.go
@@ -1,4 +1,4 @@
-package e2e
+package kube_e2e
 
 import (
 	"time"
@@ -6,17 +6,28 @@ import (
 	. "github.com/onsi/ginkgo"
 	"github.com/pborman/uuid"
 	"github.com/solo-io/gloo/pkg/api/types/v1"
+	"github.com/solo-io/gloo/pkg/plugins/kubernetes"
 	. "github.com/solo-io/gloo/test/helpers"
 )
 
-var _ = Describe("Kube Upstream Discovery", func() {
-	generatedUpstreamName := namespace + "-helloservice-8080"
-	Context("creating a vhost with a route to an upstream generated by"+
-		" kube-upstream-discovery", func() {
+var _ = Describe("Kubernetes Upstream + Endpoint Discovery Plugin", func() {
+	const helloService = "helloservice"
+	const servicePort = 8080
+	Context("creating kube upstream and a vhost with a single route to it", func() {
 		randomPath := "/" + uuid.New()
 		vhostName := "one-route"
 		BeforeEach(func() {
-			_, err := gloo.V1().VirtualHosts().Create(&v1.VirtualHost{
+			_, err := gloo.V1().Upstreams().Create(&v1.Upstream{
+				Name: helloService,
+				Type: kubernetes.UpstreamTypeKube,
+				Spec: kubernetes.EncodeUpstreamSpec(kubernetes.UpstreamSpec{
+					ServiceNamespace: namespace,
+					ServiceName:      helloService,
+					ServicePort:      servicePort,
+				}),
+			})
+			Must(err)
+			_, err = gloo.V1().VirtualHosts().Create(&v1.VirtualHost{
 				Name: vhostName,
 				Routes: []*v1.Route{{
 					Matcher: &v1.Route_RequestMatcher{
@@ -30,7 +41,7 @@ var _ = Describe("Kube Upstream Discovery", func() {
 					SingleDestination: &v1.Destination{
 						DestinationType: &v1.Destination_Upstream{
 							Upstream: &v1.UpstreamDestination{
-								Name: generatedUpstreamName,
+								Name: helloService,
 							},
 						},
 					},
@@ -39,7 +50,7 @@ var _ = Describe("Kube Upstream Discovery", func() {
 			Must(err)
 		})
 		AfterEach(func() {
-			//gloo.V1().Upstreams().Delete(generatedUpstreamName)
+			gloo.V1().Upstreams().Delete(helloService)
 			gloo.V1().VirtualHosts().Delete(vhostName)
 		})
 		It("should configure envoy with a 200 OK route (backed by helloservice)", func() {
