+++ test/kube_e2e/multi_destination_upstream_test.go
@@ -1,4 +1,4 @@
-package e2e
+package kube_e2e
 
 import (
 	"time"
@@ -6,8 +6,8 @@ import (
 	. "github.com/onsi/ginkgo"
 	"github.com/pborman/uuid"
 	"github.com/solo-io/gloo/pkg/api/types/v1"
-	. "github.com/solo-io/gloo/test/helpers"
 	"github.com/solo-io/gloo/pkg/coreplugins/service"
+	. "github.com/solo-io/gloo/test/helpers"
 )
 
 var _ = Describe("Multiple Upstream Destinations", func() {
